MT-HAL: Multi-task Highly Adaptive Lasso

Estimation procedure for MT-HAL, Multi-task Highly Adaptive Lasso

fit_mthal(
  X,
  Y,
  type,
  formula = NULL,
  max_degree = ifelse(ncol(X) >= 20, 2, 3),
  smoothness_orders = 1,
  num_knots = num_knots_generator(max_degree = max_degree, smoothness_orders =
    smoothness_orders, base_num_knots_0 = 200, base_num_knots_1 = 50),
  reduce_basis = NULL,
  lambda_1 = 10^seq(1, -10, -0.1),
  nfolds = 10,
  fit_control = list(cv_select = TRUE, prediction_bounds = "default"),
  basis_list = NULL,
  return_x_basis = FALSE,
  yolo = FALSE
)

Arguments

X

A numeric input matrix with dimensions number of observations -by- number of covariates that will be used to derive the design matrix of basis functions.

Y

A numeric matrix of observations of the outcomes with dimensions number of observations -by- number of outcomes. Missingness is permissible in this matrix, i.e., the number of observations can vary across the tasks. The outcomes can be binary or continuous. The valid value of binary outcome \(\in \{1, −1\}\).

type

The type of problem, a character that must be "Regression" or "Classification".

formula

A character string formula to be used in formula_hal. See its documentation for details.

max_degree

The highest order of interaction terms for which basis functions ought to be generated.

smoothness_orders

An integer, specifying the smoothness of the basis functions. See details for smoothness_orders for more information.

num_knots

An integer vector of length 1 or max_degree, specifying the maximum number of knot points (i.e., bins) for any covariate for generating basis functions. If num_knots is a unit-length vector, then the same num_knots are used for each degree (this is not recommended). The default settings for num_knots are recommended, and these defaults decrease num_knots with increasing max_degree and smoothness_orders, which prevents (expensive) combinatorial explosions in the number of higher-degree and higher-order basis functions generated. This allows the complexity of the optimization problem to grow scalably. See details of num_knots more information.

reduce_basis

An optional numeric value bounded in the open unit interval indicating the minimum proportion of 1's in a basis function column needed for the basis function to be included in the procedure to fit the regression. Any basis functions with a lower proportion of 1's than the cutoff will be removed. Defaults to 1 over the square root of the number of observations. Only applicable for models fit with zero-order splines, i.e. smoothness_orders = 0.

lambda_1

User-specified sequence of values of the regularization parameter for the lasso L1 regression, which control cross-task regularization. If NULL, the default sequence of 10^seq(1, -10, -.1) will be used. The cross-validated optimal value of this regularization parameter will be selected with cvMTL. If fit_control's cv_select argument is set to FALSE, then (according to the MTL documentation) "the MTL model will be trained using a warm-start technique".

nfolds

Number of V-fold cross-validation folds, default is 10. Only used when fit_control's cv_select argument is set to TRUE.

fit_control

List of arguments, including the following, and any others to be passed to cvMTL or MTL.

cv_select: A logical specifying if the sequence of specified lambda_1 values should be passed to cvMTL in order for a single, optimal value of lambda_1 to be selected according to cross-validation. When cv_select = FALSE, a MTL model will be used to fit the sequence of (or single) lambda.

prediction_bounds: An optional list of vectors of size two that provides the lower and upper bounds on the predictions for each outcome. When prediction_bounds = "default" and type = "Regression", the continuous outcome predictions will be bounded between min(y) - sd(y) and max(y) + sd(y) for each outcome y. When type = "Classification", the probability of the individual being assigned to positive label P(y==1) is estimated, and so prediction_bounds are ignored.

basis_list

The full set of basis functions generated from X.

return_x_basis

A logical indicating whether or not to return the matrix of (possibly reduced) basis functions used in fit_hal.

yolo

A logical indicating whether to print one of a curated selection of quotes from the HAL9000 computer, from the critically acclaimed epic science-fiction film "2001: A Space Odyssey" (1968).

Value

Object of class mthal9001, containing a list of basis functions, a copy map, coefficients estimated for basis functions, timing results (for assessing computational efficiency), and the RMTLfit object.

Details

The procedure uses a custom C++ implementation to generate a design matrix of spline basis functions of covariates and interactions of covariates. The regularized multivariate regression is fit to this design matrix via RMTL. The maximum dimension of the design matrix is \(n\) -by- \((n * 2^(d-1))\), where where \(n\) is the number of observations and \(d\) is the number of covariates.

For smoothness_orders = 0, only zero-order splines (piece-wise constant) are generated, which assume the true regression function has no smoothness or continuity. When smoothness_orders = 1, first-order splines (piece-wise linear) are generated, which assume continuity of the true regression function. When smoothness_orders = 2, second-order splines (piece-wise quadratic and linear terms) are generated, which assume a the true regression function has a single order of differentiability.

num_knots argument specifies the number of knot points for each covariate and for each max_degree. Fewer knot points can significantly decrease runtime, but might be overly simplistic. When considering smoothness_orders = 0, too few knot points (e.g., < 50) can significantly reduce performance. When smoothness_orders = 1 or higher, then fewer knot points (e.g., 10-30) is actually better for performance. We recommend specifying num_knots with respect to smoothness_orders, and as a vector of length max_degree with values decreasing exponentially. This prevents combinatorial explosions in the number of higher-degree basis functions generated. The default behavior of num_knots follows this logic --- for smoothness_orders = 0, num_knots is set to \(500 / 2^{j-1}\), and for smoothness_orders = 1 or higher, num_knots is set to \(200 / 2^{j-1}\), where \(j\) is the interaction degree. We also include some other suitable settings for num_knots below, all of which are less complex than default num_knots and will thus result in a faster runtime:

Some good settings for little to no cost in performance:

If smoothness_orders = 0 and max_degree = 3, num_knots = c(400, 200, 100).

If smoothness_orders = 1+ and max_degree = 3, num_knots = c(100, 75, 50).

Recommended settings for fairly fast runtime:

If smoothness_orders = 0 and max_degree = 3, num_knots = c(200, 100, 50).

If smoothness_orders = 1+ and max_degree = 3, num_knots = c(50, 25, 15).

Recommended settings for fast runtime:

If smoothness_orders = 0 and max_degree = 3, num_knots = c(100, 50, 25).

If smoothness_orders = 1+ and max_degree = 3, num_knots = c(40, 15, 10).

Recommended settings for very fast runtime:

If smoothness_orders = 0 and max_degree = 3, num_knots = c(50, 25, 10).

If smoothness_orders = 1+ and max_degree = 3, num_knots = c(25, 10, 5).

Examples

n <- 100
p <- 3
x <- xmat <- matrix(rnorm(n * p), n, p)
y_prob <- plogis(3 * sin(x[, 1]) + sin(x[, 2]))
y1 <- rbinom(n=n, size=1, prob=y_prob)
y1[y1==0] <- -1
y2 <- rbinom(n=n, size=1, prob=y_prob)
y2[y2==0] <- -1
Y <- as.matrix(cbind(y1, y2))
mthal_fit <- fit_mthal(X = x, Y = Y, type = "Classification")