Formula Interface for HAL Fitting Procedure

Formula Interface for HAL Fitting Procedure

formula_hal(
  formula,
  X,
  exclusive_dot = FALSE,
  custom_group = NULL,
  smoothness_orders = NULL,
  num_knots = NULL
)

Arguments

formula
A character string specifying the fit_hal model. The format should be of the form "~ h(x) + h(w,x) + h(x,w) + h(x,w,z)" or "y ~ h(x) + h(w,x) + h(x,w) + h(x,w,z)", where w, x, z are column names in the input matrix X and y is an (optionally-supplied) outcome. That is, the ~ is required in the formula, but anything before ~ is omitted. Each term in the formula represents main-term(s) and/or interaction(s) to be included in the model. See formula details for more information.

X
An input matrix with dimensions number of observations -by- number of covariates that will be used with the formula and other arguments defined in fit_hal to derive the design matrix of basis functions.

exclusive_dot
A logical indicator for whether the . and .^degree arguments in the formula should be treated as exclusive or with respect to the variables already specified in the formula. See details on exclusive_dot for more information. As an example, consider the formula ~ h(x,w) + .. When exclusive_dot is TRUE, the . operator adds all one-way basis functions for variables that are remaining in X and not yet specified in the formula (i.e., excluding x and w). When exclusive_dot is FALSE, the . operator adds all one-way basis functions for all variables in X (i.e., including x, w).

custom_group
A named list that represents a grouping of variables in X. Each group in the custom_group list contains a character vector of column names in X that belong to that group. The names of the custom_group must be single characters of length 1. See custom_group details for more information.

smoothness_orders
Necessary argument for generating basis functions from the formula. See its documentation in fit_hal.

num_knots
Necessary argument for generating basis functions from the formula. See its documentation in fit_hal.

Value

A formula object.

Details

This function allows users to specify the functional form/model of hal9001, similar to glm. The user can specify which interactions to include, monotonicity constraints, and smoothness constraints. The function is intended for use within fit_hal, and it is called when formula is supplied to fit_hal. This function returns a formula object, which includes parameters for subsequent use in fit_hal. In particular, HAL's formula object contains the basis_list, which is used to create the design matrix of basis functions for lasso fitting, and other parameters that are used in the lasso regression.

A formula of "h(x)" signifies that all main-term (i.e., one-way) basis functions of the variable x (in input matrix X) should be included in the model. A formula of "h(x,w)" specifies that all two-way interaction basis functions between x and w should be included in the model. Similarly, a formula of "h(x,w,z)" specifies that all three-way interaction basis functions between x, w, and z should be included in the model. Note that formula = "h(x,y,z)" will only construct three-way basis functions for x, y, and z, and not the two- and one- way basis functions. Input of the form ~ . will generate all one-term basis functions for variables in X. Similarly, ~ .^2 will generate all basis functions up to degree 2 (i.e., all one- and two- way interaction basis functions) for all variables in X. One can combine all of the notions above. For example, ~ h(x,w,z) + .^2 will generate all one- and two- way interaction basis functions for all variables in X, and additionally all the three-way interaction basis functions between variables w, x, and z.

In formula, one can also specify monotonicity constraints by replacing the letter h with d (for decreasing) or i (for increasing). For example, the formula could look like "~ i(x) + i(y) + i(x, y)", "~ d(x) + d(y) + d(x, y)", or "~ d(x) + i(y) + h(x, y)". The letters h, i, and d specify functional restrictions of each term:

h specifies no constraints on the term.

i specifies the term should be enforced to be monotonically increasing.

d specifies the term should be enforced to be monotonically decreasing.

In formula, ambiguous operations like ~ i(x) + . will use the first specification of the term in the formula (generally from left to right). That is, ~ i(x) + . is interpreted as ~ i(x) + h(z) + h(w), while ~ h(x) + i(x) is interpreted as ~ h(x). Note that . and .^degree have the lowest importance and are evaluated last, regardless of their location in the formula. As a result, ~ . + i(x) will be interpreted as ~ i(x) + h(w) + h(z), contrary to the previous case.

In formula, familiar operations such as the :, * , and - are also supported:

: is a concatenation operator which maps h(x):h(w) into h(x,w), or h(x):h(w):h(z) into h(x,w,z).

* concatenates and then generates all lower order terms/interactions. For example, h(x) * h(w) is mapped into h(x) + h(w) + h(x,w), and h(x) * h(w) * h(z) into h(x) + h(w) + h(z) + h(x,w) + h(x,z) + h(z,w) + h(x,w,z).

- subtracts/removes the term from the formula. For example, h(x) + h(w) - h(w) becomes h(x). Note that the above operations are sensitive to the constraint prefix (h, i, and d). For ambiguous operations like i(x):h(w), the unconstrained prefix h will be used unless all prefixes in the term are the same. Thus, i(x):h(w) becomes h(x,w), i(x):i(w):d(z) becomes h(x,w,z), and i(x):i(w) becomes i(x,w). The above logic is be applied recursively to *, so that something like i(x) * h(w) * i(z) is interpreted as i(x) + h(w) + i(z) + h(x,w) + i(x,z) + h(w,z) + h(x,w,z).

Another useful operation in formula is the wildcard . operator. When it's used in a specified term, it will generate all valid terms where the value of . is iterated over all variables in X. As an example, consider X with 3 columns x, w, and z. In this scenario, h(x,.) is interpreted as h(x,w) + h(x,z), h(.,.) as h(x,w) + h(x,z) + h(w,z), and h(x,w,.) as h(x,w,z). Also, in formula, all operations are compatible with one another. For example, h(.)*h(x), h(.):h(x), and h(x) - h(.) are all valid and behave as expected.

Note that, if exclusive_dot is FALSE, then differently-appearing formulas can actually be identical. For example, ~ h(x) + .^2 and ~ .^2 specify the same formula (i.e., generating all basis functions up to degree 2) when exclusive_dot is FALSE. However, if exclusive_dot was set to TRUE, then ~ h(x) + .^2 encodes a different formula than ~ .^2; specifically, ~ h(x) + .^2 means to generate one-way basis functions for the variable x, and all basis functions up to degree 2 for all other variables in X (i.e., excluding x).

The custom_group allows the user to specify their own wildcard symbols (e.g., .); however, the value of the symbol will be iterated over all variables specified in the user-supplied group. For example, if one sets custom_group = list("group1" = c("x", "w"), "group2" = c("t","r")), then the formula would be mapped from ~ h(group1, group2) to ~ h(x,t) + h(x,r) + h(w,t) + h(w,r), so that all two-way interactions using one variable from each group are generated. Similarly, ~ h(1,r) would be mapped to ~ h(x,r) + h(w,r). Thus, the custom groups operate exactly as ., except the possible values are restricted to a specific group.