Prediction from HAL fits

Prediction from HAL fits

# S3 method for hal9001
predict(
  object,
  offset = NULL,
  lambda = NULL,
  ...,
  new_data,
  new_X_unpenalized = NULL
)

Arguments

object
An object of class hal9001, containing the results of fitting the Highly Adaptive Lasso, as produced by a call to fit_hal.

offset
A vector of offsets. Must be provided if provided at training

lambda
A single lambda value or a vector of lambdas to use for prediction. If NULL, a value of lambda will be selected based on cross-validation, using cv.glmnet.

...
Additional arguments passed to predict as necessary.

new_data
A matrix or data.frame containing new data (observations NOT used in fitting the hal9001 object passed in via the object argument above) for which the hal9001 object will compute predicted values.

new_X_unpenalized
If the user supplied X_unpenalized during training, the user should also supply this matrix with the same number of observations as new_data. Optional.

Value

A numeric vector of predictions from a fitted hal9001 object.

Details

Method for computing and extracting predictions from fits of the Highly Adaptive Lasso estimator, returned as a single S3 objects of class hal9001.