Screen HAL Columns, Basis Functions, and lambda

Smart Screening Stuff. TODO: Document fully

hal_screen_cols(
  x,
  y,
  V = 5,
  family,
  col_lists = NULL,
  foldid = NULL,
  offset = NULL,
  verbose = FALSE
)

hal_screen_basis(
  x,
  y,
  family,
  foldid = NULL,
  offset = NULL,
  verbose = FALSE,
  max_degree = NULL,
  main_terms = NULL
)

hal_screen_lambda(
  x_basis,
  y,
  family,
  offset = NULL,
  foldid = NULL,
  lambda = NULL
)

Arguments

x
An input matrix containing observations of covariates.

y
A numeric vector of obervations of the outcome variable.

V
A numeric of the number of folds to use in cross-validation. Defaults to five. If foldid is not specified, this is used.

family
A character corresponding to the error family for a generalized linear model. Options are limited to "gaussian" for fitting a standard general linear model and "binomial" for logistic regression.

col_lists
A list of lists of column number, indicating which basis columns to screen.

foldid
A vector of fold IDs, as in cv.glmnet.

offset
A vector of offset values, used in fitting.

verbose
If TRUE, print details of screening steps.

max_degree
The highest order of interaction terms for which the basis functions ought to be generated. The default (NULL) corresponds to generating basis functions for the full dimensionality of the input matrix.

main_terms
If TRUE, only screen interactions for siginficant main terms

x_basis
An x_basis sparse matrix.

lambda
A user-specified array of values of the lambda tuning parameter of the Lasso L1 regression. cv.glmnet will be used when set to NULL, automatically selecting an optimal value based on a cross-validated fit criterion (e.g., MSE). If specified, Lasso L1 regression model will be fit via glmnet, returning regularized coefficient values for each value in the input array.