Wrapper for Classic SuperLearner

Wrapper for SuperLearner for objects of class hal9001

SL.hal9001(
  Y,
  X,
  newX = NULL,
  family = stats::gaussian(),
  obsWeights = rep(1, length(Y)),
  id = NULL,
  ...
)

Arguments

Y
A numeric vector of observations of the outcome variable.

X
A matrix of predictors/covariates.

newX
A matrix of new observations on which to obtain predictions. The default of NULL computes predictions on training inputs X.

family
A family object (one that is supported by glmnet) specifying the error/link family for a generalized linear model.

obsWeights
A numeric vector of observational-level weights.

id
A numeric vector of IDs.

...
Additional arguments passed to fit_hal. See its documentation for more details.

Value

An object of class SL.hal9001 with a fitted hal9001 object and corresponding predictions based on the input data.