Wrapper for Classic SuperLearner

Wrapper for package SuperLearner for objects of class hal9001

SL.hal9001(Y, X, newX = NULL, degrees = NULL, fit_type = c("glmnet",
  "lassi"), n_folds = 10, use_min = TRUE, family = stats::gaussian(),
  obsWeights = rep(1, length(Y)), ...)

Arguments

Y
A numeric of outcomes.

X
A matrix of predictors/covariates.

newX
A matrix of new observations on which to obtain predictions. The default of NULL computes predictions on training inputs X.

degrees
The highest order of interaction terms for which the basis functions ought to be generated. The default (NULL) corresponds to generating basis functions for the full dimensionality of the input matrix.

fit_type
The specific routine to be called when fitting the LASSO regression in a cross-validated manner. Choosing the glmnet option will result in a call to cv.glmnet while lassi will produce a (faster) call to a custom routine based on the lassi package.

n_folds
Integer for the number of folds to be used when splitting the data for cross-validation. This defaults to 10 as this is the convention for v-fold cross-validation.

use_min
Determines which lambda is selected from cv.glmnet. TRUE corresponds to "lambda.min" and FALSE corresponds to "lambda.1se".

family
Not used by the function directly, but meant to ensure compatibility with SuperLearner.

obsWeights
Not used by the function directly, but meant to ensure compatibility with SuperLearner. These are passed to cv.glmnet or glmnet through the ... argument of fit_hal.

...
Prevents process death. DON'T USE.