Formula Interface for HAL Fitting Procedure

Estimation procedure for HAL, the Highly Adaptive Lasso

formula_hal(
  formula,
  data,
  smoothness_orders = NULL,
  num_knots = NULL,
  exclusive_dot = FALSE,
  custom_group = NULL,
  adaptive_smoothing = FALSE,
  ...
)

fit_hal_formula(formula, ...)

fit_hal(
  X,
  Y,
  X_unpenalized = NULL,
  max_degree = ifelse(ncol(X) >= 20, 2, 3),
  smoothness_orders = rep(1, ncol(X)),
  num_knots = sapply(seq_len(max_degree), num_knots_generator, smoothness_orders =
    smoothness_orders, base_num_knots_0 = 500, base_num_knots_1 = 200),
  fit_type = c("glmnet", "lassi"),
  n_folds = 10,
  foldid = NULL,
  use_min = TRUE,
  reduce_basis = NULL,
  family = c("gaussian", "binomial", "poisson", "cox"),
  return_lasso = TRUE,
  return_x_basis = FALSE,
  basis_list = NULL,
  lambda = NULL,
  id = NULL,
  offset = NULL,
  cv_select = TRUE,
  adaptive_smoothing = FALSE,
  prediction_bounds = "default",
  p_reserve = 0.5,
  ...,
  yolo = FALSE
)

Arguments

formula
A character string specifying the hal9001 model. The format should be of the form "y ~ h(x) + h(w,x) + h(x,w) + h(x,w,z) ", where "y" is the outcome and "w,x,y,z" are variables in data. Each term represents a main-term/interaction to be included in the model. h(x) signifies that all one-way/main terms basis functions of the variable x should be included. h(x,w) specifies that all interaction (two-way) basis functions between x and w should be included in the model. Similarly, h(x,w,z) specifies that all interaction (three-way) basis functions between x,w,z should be included in the model. Note that "y ~ h(x,y,z)" will only construct three-way basis functions for x, y, z and not the two-way and one-way basis functions. Additionally, a formula of the form "y ~ ." will generate all one-way main term basis functions for variables in data. Similarly, "y ~ .^2" will generate all basis functions up to degree 2 for all variables in data. More generally, "y ~ .^max_degree" will construct all basis functions up to degree max_degree. One can combine all the notions above. For example, "y ~ h(x,w,z) + ." and "y ~ h(x,w,z) + .^2" will generate all one-way (respectively, up to two-way) basis functions and additionally all the three-way interaction basis functions between variables w, x, z. One can also specify monotonicity constraints by replacing the letter h with d (for decreasing) or i (for increasing), e.g., formulas like "y ~ i(x) + i(y) + i(x, y)", "y ~ d(x) + d(y) + d(x, y)", or "y ~ d(x) + i(y) + h(x, y)". The letters h, i, d specify functional restrictions of each term:

h specifies no constraints on the term,

i specifies the term should be enforced to be monotonely increasing,

d specifies the term should be enforced to be monotonely decreasing. Ambiguous operations like "y ~ i(x) + ." will use the first specification of the term in the formula (generally from left to right). That is, "y ~ i(x) + ." is interpreted as "y ~ i(x) + h(z) + h(w)" while "y ~ h(x) + i(x)" is interpreted as "y ~ h(x)". NOTE that "." and ".^max_degree" have the lowest importance and are evaluated last, regardless of their location in the formula. As a result, "y ~ . + i(x)" will be interpreted as "y ~ i(x) + h(w) + h(z)", contrary to the previous case. Familiar operations such as the :, * , - are also supported:

: is a concatnation operator which maps h(x):h(w) into h(x,w) or h(x):h(w):h(z) into h(x,w,z).

* concatenates and then generates all lower order terms/interactions. For example, h(x)*h(w) into h(x) + h(w) + h(x,w) or h(x)*h(w)*h(z) into h(x) + h(w) + h(z) + h(x,w) + h(x,z) + h(z,w) + h(x,w,z).

- subtracts/removes the term from the formula. For example, h(x) + h(w) - h(w) becomes h(x). Note that the above operations are sensitive to the constraint prefix (h, i, and d). For ambigious operations like i(x):h(w), the unconstrained prefix h will be used unless all prefixes in the term are the same. So, i(x):h(w) becomes h(x,w) and i(x):i(w):d(z) becomes h(x,w,z), and i(x):i(w) becomes i(x,w). The above logic will be applied recursively to * so that i(x) * h(w) * i(z) is interpreted as i(x) + h(w) + i(z) + h(x,w) + i(x,z) + h(w,z) + h(x,w,z). Another useful operation is the wildcard . operator, which when used in a specified term will generate all valid terms where the value of . is iterated over the non-outcome columns in the data matrix. For example, h(x,.) is h(x,w) + h(x,z) and h(.,.) is h(x,w) + h(x,z) + h(w,z), and h(x,w,. is h(x,w,z), assuming the covariates are only (x, w, z). All operations are compatible with one another, e.g., h(.)*h(x), h(.):h(x) and h(x) - h(.) are valid and behave as expected.

data
A data.frame or named matrix containing the outcome and covariates specified in the argument formula.

smoothness_orders
An integer vector of length 1 or length ncol(X). If smoothness_orders is of length 1, then its values are recycled to form a vector of length ncol(X). Given such a vector of length ncol(X), the ith element specifies the level of smoothness for the variable corresponding with the ith column in X. A value of "0" corresponds with 0th-order splines (piece-wise constant), which assumes no smoothness or continuity of the true regression function. A value of "1" corresponds with 1st-order splines (piece-wise linear), which only assumes continuity of the true regression function. A value of "2" corresponds with 2nd-order splines (piece-wise quadratic and linear terms), which assumes a single order of differentiability for the true regression function. NOTE: If smoothness_orders has length less than ncol(X), then its values are recycled as needed.

num_knots
An integer vector of length 1 or length max_degree. If num_knots is a vector of length 1 then its values are recycled to produce a vector of length max_degree. Given a possibly recycled vector of length max_degree, num_knots[i] specifies the maximum number of knot points used when generating basis functions of degree i for each covariate. For example, num_knots[1] specifies how many knot points to use when generating main-term additive basis functions. num_knots[2] specifies how many knot points should be used when generating each univariate basis function in the 2-tensor product basis functions. A smaller number of knot points gives rise to a less smooth function. However, fewer knot points can significantly decrease runtime. If smoothness_orders is 1 or higher then few knot points (10-30) are needed to maintain near-optimal performance. When considering setting smoothness_orders = 0, too few knot points (<50) can significantly reduce performance; thus, we recommend specifying a vector of length max_degree that decreases exponentially, preventing combinatorial explosions in the number of higher-degree basis functions generated. Default: For zero order smoothness (any(smoothness_orders==0)), the number of knots by interaction degree d decays as \(500/2^{d-1}\). For first or higher-order smoothness (all(smoothness_orders>0)), the number of knots by interaction degree d decays as \(75/2^{d-1}\). These defaults ensure that the number of basis functions and thus the complexity of the optimization problem grows scalably in max_degree.

Some good settings for little to no cost in performance:

If smoothness_orders = 0, max_degree = 3, num_knots = c(400, 200, 100).

If smoothness_orders = 1+, max_degree = 3, num_knots = c(100, 75, 50).

Recommended settings for fairly fast runtime and great performance:

If smoothness_orders = 0, max_degree = 3, num_knots = c(200, 100, 50).

If smoothness_orders = 1+, max_degree = 3, num_knots = c(50, 25, 15).

Recommended settings for fast runtime and good/great performance:

If smoothness_orders = 0, max_degree = 3, num_knots = c(100, 50, 25).

If smoothness_orders = 1+, max_degree = 3, num_knots = c(40, 15, 10).

Recommended settings for very fast runtime and good performance:

If smoothness_orders = 0, max_degree = 3, num_knots = c(50, 25, 10).

If smoothness_orders = 1+, max_degree = 3, num_knots = c(25, 10, 5).

exclusive_dot
A logical indicator for whether the ``. and .^max_degree arguments in the formula should be treated as exclusive or inclusive with respect to the variables already specified in the formula. For example, in `y ~ h(x,w) + .`, should the `.` operator be interpreted as

add all one-way basis functions for variables remaining in data not yet specified in the formula (i.e., excluding x, w); or,

add all one-way basis functions for all variables in the data (including x, w). As an example, if exclusive_dot is set to FALSE, then y ~ h(x) + .^2 and y ~ .^2 specify the same formula, i.e., generating all basis functions up to degree 2; however, if exclusive_dot is set to TRUE, then y ~ h(x) + .^2 encodes a different formula than y ~ .^2. Specifically, it means to generate one-way basis functions for the variable "x" and then all basis functions up to degree 2 for other variables excluding "x" in data. As a result, no interactions will be added for the variable "x".

custom_group
A named list with single character names that represent a group, with its elements being a character vector of variable names. This allows the user to specify their own wildcard symbols (e.g., .); however, the value of the symbol will be iterated over all variables specified in the user-supplied group. For example, if one sets custom_group = list("1" = c("x", "w"), "2" = c("t","r")), then the following formula is mapped from y ~ h(1,2) to y ~ h(x,t) + h(x,r) + h(w,t) + h(w,r), so that all two-way interactions using one variable for each group are generated. Similarly, y ~ h(1,r) would be mapped to y ~ h(x,r) + h(w,r). Thus, the custom groups operate exactly as ., except the possible values are restricted to a specific group.

adaptive_smoothing
A logical, which, if TRUE, HAL will perform adaptive smoothing up until the maximum order of smoothness specified by smoothness_orders. For example, if smoothness_orders = 2 and adaptive_smoothing = TRUE, then HAL will generate all basis functions of smoothness order 0, 1, and 2, and data-adaptively select the basis functions to use. WARNING: This can increase runtime by a factor of 2-3 times depending on value of smoothness_orders.

...
Other arguments passed to cv.glmnet. Please consult its documentation for a full list of options.

X
An input matrix containing observations and covariates.

Y
A numeric vector of obervations of the outcome variable.

X_unpenalized
An input matrix with the same format as X, that directly get appended into the design matrix (no basis expansion). No L1 penalization is performed on these covariates.

max_degree
The highest order of interaction terms for which the basis functions ought to be generated. The default (NULL) corresponds to generating basis functions for the full dimensionality of the input matrix.

fit_type
The specific routine to be called when fitting the Lasso regression in a cross-validated manner. Choosing the glmnet option will result in a call to cv.glmnet while lassi will produce a (faster) call to a custom Lasso routine.

n_folds
Integer for the number of folds to be used when splitting the data for V-fold cross-validation. This defaults to 10.

foldid
An optional numeric containing values between 1 and n_folds, identifying the fold to which each observation is assigned. If supplied, n_folds can be missing. In such a case, this vector is passed directly to cv.glmnet.

use_min
Specify lambda selected by cv.glmnet. TRUE, "lambda.min" is used; otherwise, "lambda.1se".

reduce_basis
A numeric value bounded in the open unit interval indicating the minimum proportion of 1's in a basis function column needed for the basis function to be included in the procedure to fit the Lasso. Any basis functions with a lower proportion of 1's than the cutoff will be removed. This argument defaults to NULL, in which case all basis functions are used in the lasso-fitting stage of the HAL algorithm.

family
A character or a family object (supported by glmnet) corresponding to the error/link family for a generalized linear model. character options are limited to "gaussian" for fitting a standard penalized linear model, "binomial" for penalized logistic regression, "poisson" for penalized Poisson regression, and "cox" for a penalized proportional hazards model. Note that in all cases where family is not set to "gaussian", fit_type is limited to "glmnet". NOTE: Passing in family objects lead to signficantly slower performance relative to passing in a character family (if supported). Thus, for nonparametric (HAL) logistic regression, one should always set family = "binomial" and never set family = binomial().

return_lasso
A logical indicating whether or not to return the glmnet fit of the lasso model.

return_x_basis
A logical indicating whether or not to return the matrix of (possibly reduced) basis functions used in the HAL lasso fit.

basis_list
The full set of basis functions generated from the input data X (via a call to enumerate_basis). The dimensionality of this structure is dim = \((n * 2^(d-1))\), where n is the number of observations and d is the number of columns in X.

lambda
User-specified array of values of the lambda tuning parameter of the Lasso L1 regression. If NULL, cv.glmnet will be used to automatically select a CV-optimal value of this regularization parameter. If specified, the Lasso L1 regression model will be fit via glmnet, returning regularized coefficient values for each value in the input array.

id
a vector of ID values, used to generate cross-validation folds for cross-validated selection of the regularization parameter lambda.

offset
a vector of offset values, used in fitting.

cv_select
A logical specifying whether the array of values specified should be passed to cv.glmnet in order to pick the optimal value (based on cross-validation) (when set to TRUE) or to simply fit along the sequence of values (or single value) using glmnet (when set to FALSE).

prediction_bounds
A vector of size two that provides the lower and upper bounds for predictions. By default, the predictions are bounded between min(Y) - sd(Y) and max(Y) + sd(Y). Bounding ensures that there is no extrapolation and that predictions remain bounded, which is is necessary for cross-validation selection and/or Super Learning.

p_reserve
Sparse matrix pre-allocation proportion. Default value is 0.5. If a dense HAL design matrix is expected, it would be useful to set p_reserve to a higher value.

yolo
A logical indicating whether to print one of a curated selection of quotes from the HAL9000 computer, from the critically acclaimed epic science-fiction film "2001: A Space Odyssey" (1968).

Value

Object of class hal9001, containing a list of basis functions, a copy map, coefficients estimated for basis functions, and timing results (for assessing computational efficiency).

Details

The function allows users to specify the functional form/model of hal9001 similar to in glm. The user can specify which interactions to include, monotonicity constraints, and smoothness constraints. The returned formula object can be fed directly into fit_hal and the fit can be run with minimal (no) user input.

The procedure uses a custom C++ implementation to generate a design matrix consisting of basis functions corresponding to covariates and interactions of covariates and to remove duplicate columns of indicators. The Lasso regression is fit to this (usually) very wide matrix using either a custom implementation (based on origami) or by a call to cv.glmnet.

Examples

# \donttest{
n <- 100
p <- 3
x <- xmat <- matrix(rnorm(n * p), n, p)
y_prob <- plogis(3 * sin(x[, 1]) + sin(x[, 2]))
y <- rbinom(n = n, size = 1, prob = y_prob)
ml_hal_fit <- fit_hal(X = x, Y = y, family = "binomial", yolo = FALSE)
preds <- predict(ml_hal_fit, new_data = x)
# }