HAL: The Highly Adaptive LASSO estimator

Estimation procedure for HAL, the Highly Adaptive LASSO

fit_hal(X, Y, degrees = NULL, fit_type = c("origami", "glmnet"),
  n_folds = 10, use_min = TRUE, ..., yolo = TRUE)

Arguments

X
An input matrix containing observations and covariates following standard conventions in problems of statistical learning.

Y
A numeric vector of obervations of the outcome variable of interest, following standard conventions in problems of statistical learning.

degrees
The highest order of interaction terms for which the basis functions ought to be generated. The default (NULL) corresponds to generating basis functions for the full dimensionality of the input matrix.

fit_type
The specific routine to be called when fitting the LASSO regression in a cross-validated manner. Choosing the glmnet option will result in a call to cv.glmnet while origami will produce a (faster) call to a custom routine based on the origami package.

n_folds
Integer for the number of folds to be used when splitting the data for cross-validation. This defaults to 10 as this is the convention for v-fold cross-validation.

use_min
Determines which lambda is selected from cv.glmnet. TRUE corresponds to "lambda.min" and FALSE corresponds to "lambda.1se".

...
Other arguments passed to cv.glmnet. Please consult the documentation for glmnet for a full list of options.

yolo
A logical indicating whether to print one of a curated selection of quotes from the HAL9000 computer, from the critically acclaimed epic science-fiction film "2001: A Space Odyssey" (1968).

Value

Object of class hal9001, containing a list of basis functions, a copy map, coefficients estimated for basis functions, and timing results (for assessing computational efficiency).

Details

The procedure uses a custom C++ implementation to generate a design matrix (consisting of basis functions corresponding to covariates and interactions of covariates) and remove duplicate columns of indicators. The LASSO regression is fit to this (usually) very wide matrix using either a custom implementation (based on the origami package) or by a call to cv.glmnet from the glmnet package.