Single LASSO estimation for cross-validation with Origami

Fits the LASSO regression over a single fold of a cross-validated data set. This is meant to be called using cross_validate from the origami package, which is done through cv_lasso. Note that this procedure is NOT meant to be invoked by itself. INTERNAL USE ONLY.

lassi_origami(fold, data, lambdas)

Arguments

fold
A fold object produced by a call to make_folds from the origami package.

data
A dgCMatrix object containing the outcome values (Y) in its first column and vectors corresponding to the basis functions of HAL in all other columns. Consult the description of the HAL algorithm for details.

lambdas
A numeric vector corresponding to a sequence of lambda values obtained by fitting the LASSO on the full data.